requirements.txt
streamlit
google-generativeai  
python-dotenv
langchain
PyPDF2
faiss-cpu  #vector store
langchain_google_genai  #using langchain asses api s of google genai